pyproject.toml
[project]
name = "ai-agent-development"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.62.0",
    "chromadb>=1.0.15",
    "docling>=2.48.0",
    "google-genai>=1.9.0",
    "gradio>=5.42.0",
    "langchain>=0.3.23",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.21",
    "langchain-core>=0.3.51",
    "langchain-google-genai>=2.1.2",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.6",
    "langgraph-cli[inmem]>=0.4.2",
    "loguru>=0.7.3",
    "notebook>=7.3.3",
    "openai>=1.98.0",
    "openai-agents>=0.2.4",
    "pymupdf>=1.26.3",
    "pypdf>=5.9.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "torch>=2.7.1",
    "xai-sdk>=1.0.1",
]
#uv sync upgrade # to update all dependencies